USA Executions Since Gregg v. Georgia (1976)

This dataset contains a list of executions in the United States since the 1976 Supreme Court decision in Gregg v. Georgia, which reinstated the death penalty in the U.S. The dataset includes details about the victims, perpetrators, and execution methods used. It is a cleaned and abbreviated version of the original dataset, which can be obtained at the source link below.

Usage

executions

Format

executions

A data frame with 2332 rows and 12 columns:

victim_id

Unique ID of a victim of an individual who was executed

victim_race

The race of the victim

victim_sex

The sex of the victim

execution_id

Unique ID of the individual who was executed, the perpetrator

perpetrator_name

Name of the perpetrator

perpetrator_race

Race of the perpetrator

perpetrator_sex

Sex of the perpetrator

perpetrator_age

Age of the perpetrator

state_of_execution

State in which the execution occurred

federal_execution_flag

Whether or not the execution was federal

execution_date

Date that the execution occurred

execution_method

Method used for the execution

Source

https://deathpenaltyinfo.org/facts-and-research/data/executions