Shooting Incidents in New York City (2006-2024)

This dataset provides a breakdown of every shooting incident that occurred in NYC from 2006 through the end of the previous calendar year. The data is manually extracted every quarter and reviewed by the Office of Management Analysis and Planning before being posted on the NYPD website. Each record represents a shooting incident in NYC and includes information about the event, the location, and the time of occurrence. Additionally, information related to suspect and victim demographics is also included. This dataset is a cleaned and abbreviated version of the original NYPD shooting incident data, which can be obtained at the source link below.

Usage

ny_shooting

Format

ny_shooting

A data frame with 23298 rows and 17 columns:

incident_id

Unique ID of the shooting incident

occur_date

Date of the shooting incident

occur_time

Time of the shooting incident

borough

Borough in which the shooting incident occurred

precinct

Identifier of the precinct of the shooting incident

location_class

Classification of the location where the shooting incident occurred

location_description

Description of the location where the shooting incident occurred

indoor_flag

Indicator of whether the shooting occurred indoors

latitude

Latitude of the location of the shooting incident

longitude

Longitude of the location of the shooting incient

victim_age_group

Age group of the victim involved the shooting incident

victim_sex

Sex of the victim involved in the shooting incident

victim_race_ethnicity

Race/ethnicity of the victim involved in the shooting incident

perpetrator_age_group

Age group of the perpetrator

perpetrator_sex

Sex of the perpetrator

perpetrator_race_eth

Race/ethnicity of the perpetrator

murder_flag

Indicator of whether the shooting incident was a murder

Source

https://catalog.data.gov/dataset/nypd-shooting-incident-data-historic